refactor(ComboBox): moved `focusNextElement` to Effects
